Reduce the content of entry in phonebook_opt.h
Added another new struct 'detailEntry', which include the data that
was in entry before (firstName, phone, and addrress etc), except
lastName. Leave lastName and pNext and 'detail' pointer which is
'detailEntry' type in entry. It make the size of entry left only 32
bytes only. By benefit of it, we can put much more entry in cache,
and it could make cache-miss rate lower obviously. The time cost to
append() and findName() is also less than before.

However, the enhance is only in search speed's point of view. Do not
consider about the storage space of detailEntry. Comparing to the
version before, it use more total memory space to storage whole
phonebook.